# docs/swagger.yaml
basePath: /
definitions:
  http.cancelRequest:
    properties:
      provider_transaction_id:
        type: string
    required:
    - provider_transaction_id
    type: object
  http.depositRequest:
    properties:
      amount:
        type: number
      currency:
        type: string
      provider_transaction_id:
        type: string
      provider_withdrawn_transaction_id:
        type: string
    required:
    - currency
    - provider_transaction_id
    - provider_withdrawn_transaction_id
    type: object
  http.loginRequest:
    properties:
      password:
        example: testpass
        type: string
      username:
        example: testuser1
        type: string
    required:
    - password
    - username
    type: object
  http.withdrawRequest:
    properties:
      amount:
        type: number
      currency:
        type: string
      game_id:
        type: string
      provider_transaction_id:
        type: string
      round_id:
        type: string
    required:
    - amount
    - currency
    - provider_transaction_id
    type: object
host: localhost:8080
info:
  contact: {}
  description: A Game Integration API for casino games with wallet management. Provides
    authentication, player information, bet placement (withdraw), bet settlement (deposit),
    and transaction cancellation endpoints.
  title: Game Integration API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token and username
      parameters:
      - description: User credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/http.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{token: string, username: string}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties: true
            type: object
      summary: Authenticate user
      tags:
      - Auth
  /bet/cancel:
    post:
      consumes:
      - application/json
      description: Cancel a bet transaction
      parameters:
      - description: Cancel details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.cancelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Bet response
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Cancel a transaction
      tags:
      - Bet
  /bet/deposit:
    post:
      consumes:
      - application/json
      description: Settle a bet by depositing funds
      parameters:
      - description: Deposit details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.depositRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Bet response
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Settle a bet (deposit)
      tags:
      - Bet
  /bet/withdraw:
    post:
      consumes:
      - application/json
      description: Place a bet by withdrawing funds
      parameters:
      - description: Withdraw details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.withdrawRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Bet response
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Place a bet (withdraw)
      tags:
      - Bet
  /metrics:
    get:
      description: Get application metrics in Prometheus format.
      produces:
      - text/plain
      responses:
        "200":
          description: Prometheus metrics
          schema:
            type: string
      summary: Get application metrics
      tags:
      - Metrics
  /profile:
    get:
      description: Get the authenticated player's profile
      produces:
      - application/json
      responses:
        "200":
          description: Profile response
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get player profile
      tags:
      - Player
securityDefinitions:
  BearerAuth:
    description: 'IMPORTANT: Enter your JWT token with "Bearer " prefix. Example:
      Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoyOSwiZXhwIjoxNzUxMjg4ODc0fQ.IwLr7sPvhXb_3HxI4d8F_UQinvJxc3ePfuM30ztMcdU'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"